+++ packages/eslint-plugin-jsx/src/rules/no-spreading-key.ts
@@ -13,7 +13,7 @@ export type MessageID = ConstantCase<typeof RULE_NAME>;
 export default createRule<[], MessageID>({
   name: RULE_NAME,
   meta: {
-    type: "suggestion",
+    type: "problem",
     docs: {
       description: "disallow spreading `key` from objects.",
       recommended: "recommended",