+++ packages/eslint-plugin-jsx/src/rules/prefer-shorthand-fragment.ts
@@ -11,7 +11,7 @@ export type MessageID = ConstantCase<typeof RULE_NAME>;
 export default createRule<[], MessageID>({
   name: RULE_NAME,
   meta: {
-    type: "suggestion",
+    type: "problem",
     docs: {
       description: "enforce using fragment syntax instead of `Fragment` component",
       recommended: "recommended",