+++ packages/eslint-plugin-naming-convention/src/rules/filename-extension.ts
@@ -62,7 +62,7 @@ const schema = [
 export default createRule<Options, MessageID>({
   name: RULE_NAME,
   meta: {
-    type: "suggestion",
+    type: "problem",
     docs: {
       description: "enforce naming convention for JSX file extensions",
       requiresTypeChecking: false,