+++ packages/eslint-plugin-react/src/rules/no-children-prop.ts
@@ -41,7 +41,7 @@ export type MessageID = ConstantCase<typeof RULE_NAME>;
 export default createRule<[], MessageID>({
   name: RULE_NAME,
   meta: {
-    type: "suggestion",
+    type: "problem",
     docs: {
       description: "disallow passing of `children` as props",
       recommended: "recommended",